fix(pomodoro): fix config defaults and deduplicate chime logic
- move autoStartBreak/autoStartWork from pomodoroConfig to focusModeConfig in postgres and sqlite DB defaults to match frontend types
- align focusModeConfig PiP default sizes with frontend (120x100 instead of 240x220)
- strip migrated autoStartBreak/autoStartWork from pomodoroConfig in SettingsContext migration to prevent re-migration cycle
- extract shared playChime function to src/lib/audio-chime.ts, replace duplicates in usePomodoroTimer and useTravelerTimer
- fix HttpApiPersistence.listPomodoroSessions to properly unwrap { data } wrapper for consistent return shape
- add destroyTravelerIdleState cleanup method to traveler-idle-state module
