+++ server/db/postgres.ts
@@ -153,17 +153,17 @@ const DEFAULT_APP_SETTINGS: AppSettingsEntity = {
     breakDuration: 5 * 60 * 1000,
     longBreakDuration: 15 * 60 * 1000,
     cyclesBeforeLongBreak: 4,
-    autoStartWork: false,
-    autoStartBreak: true,
     pomodoroEnabled: false,
   },
   focusModeConfig: {
     enablePiP: true,
-    pipWidth: 240,
-    pipHeight: 220,
+    pipWidth: 120,
+    pipHeight: 100,
     wakeLock: true,
     soundNotifications: true,
     showFocusOverlay: false,
+    autoStartBreak: true,
+    autoStartWork: false,
   },
   travelerConfig: {
     travelMode: 'flight',
