+++ src/hooks/usePomodoroTimer.ts
@@ -4,42 +4,7 @@ import { useSettingsContext } from '@/context/SettingsContext';
 import { useCurrentUser } from '@/hooks/useCurrentUser';
 import { getPersistenceAdapter } from '@/lib/persistence-factory';
 import { eventBus } from '@/lib/events';
-
-let audioCtx: AudioContext | null = null;
-
-function playChime(type: 'work' | 'break'): void {
-  try {
-    if (!audioCtx) {
-      audioCtx = new AudioContext();
-    }
-    if (audioCtx.state === 'suspended') {
-      audioCtx.resume();
-    }
-    const ctx = audioCtx;
-    const osc = ctx.createOscillator();
-    const gain = ctx.createGain();
-    osc.connect(gain);
-    gain.connect(ctx.destination);
-
-    if (type === 'work') {
-      osc.frequency.setValueAtTime(523.25, ctx.currentTime);
-      osc.frequency.setValueAtTime(659.25, ctx.currentTime + 0.15);
-      osc.frequency.setValueAtTime(783.99, ctx.currentTime + 0.3);
-    } else {
-      osc.frequency.setValueAtTime(783.99, ctx.currentTime);
-      osc.frequency.setValueAtTime(659.25, ctx.currentTime + 0.15);
-      osc.frequency.setValueAtTime(523.25, ctx.currentTime + 0.3);
-    }
-
-    osc.type = 'sine';
-    gain.gain.setValueAtTime(0.3, ctx.currentTime);
-    gain.gain.exponentialRampToValueAtTime(0.01, ctx.currentTime + 0.6);
-    osc.start(ctx.currentTime);
-    osc.stop(ctx.currentTime + 0.6);
-  } catch {
-    // AudioContext may not be available
-  }
-}
+import { playChime } from '@/lib/audio-chime';
 
 const INITIAL_STATE: PomodoroState = {
   phase: 'idle',
